Fragments refactored
