+++ caesars-palace/blocks/fragment/fragment.js
@@ -1,29 +1,6 @@
 import {
-  decorateMain,
+  loadFragment,
 } from '../../scripts/scripts.js';
-
-import {
-  loadBlocks,
-} from '../../scripts/lib-franklin.js';
-
-/**
- * Loads a fragment.
- * @param {string} path The path to the fragment
- * @returns {HTMLElement} The root element of the fragment
- */
-async function loadFragment(path) {
-  if (path && path.startsWith('/')) {
-    const resp = await fetch(`${path}.plain.html`);
-    if (resp.ok) {
-      const main = document.createElement('main');
-      main.innerHTML = await resp.text();
-      decorateMain(main);
-      await loadBlocks(main);
-      return main;
-    }
-  }
-  return null;
-}
 
 export default async function decorate(block) {
   const link = block.querySelector('a');
